requirements.txt
pymongo==4.6.1
neo4j==5.16.0
python-dotenv==1.0.0
matplotlib==3.8.2
pyvis==0.3.2
plotly>=5.18.0
streamlit>=1.30.0
textblob>=0.17.1